refactor: move ParametricExpression overloads to separate file
