+++ src/SymbolicRegression.jl
@@ -217,6 +217,7 @@ using DispatchDoctor: @stable
     include("Migration.jl")
     include("SearchUtils.jl")
     include("ExpressionBuilder.jl")
+    include("ParametricExpression.jl")
 end
 
 using .CoreModule:
